Move the positions of some tests
