+++ src/test/java/seedu/address/model/property/AddressTest.java
@@ -1,4 +1,4 @@
-package seedu.address.model.person;
+package seedu.address.model.property;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
